requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.11.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.1
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.8.0
Django==4.1.3
django-celery-beat==2.4.0
django-environ==0.9.0
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-phonenumber-field==7.0.0
django-timezone-field==5.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
flower==1.2.0
fonttools==4.38.0
fpdf2==2.6.0
humanize==4.4.0
idna==3.4
importlib-metadata==5.0.0
jsonschema==3.2.0
kombu==5.2.4
Markdown==3.4.1
memory-profiler==0.61.0
netaddr==0.8.0
packaging==23.0
paramiko==3.0.0
phonenumberslite==8.13.0
Pillow==9.3.0
pipreqs==0.4.11
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.6
PyYAML==5.4.1
redis==4.4.1
reportlab==3.6.12
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
texttable==1.6.7
tornado==6.2
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.59.0
wrapt==1.14.1
yarg==0.1.9
yookassa==2.3.3
zipp==3.10.0
